refactor: add reporting module to framework
