+++ gateway-framework/Cargo.toml
@@ -8,6 +8,7 @@ alloy-primitives.workspace = true
 alloy-sol-types.workspace = true
 anyhow.workspace = true
 axum.workspace = true
+chrono = { version = "0.4", default-features = false, features = ["clock"] }
 candidate-selection = { git = "https://github.com/edgeandnode/candidate-selection", rev = "bcd7909" }
 ethers = "2.0.14"
 eventuals = "0.6.7"
@@ -22,6 +23,7 @@ ordered-float = "4.2.0"
 primitive-types.workspace = true
 prometheus = "0.13.3"
 rand.workspace = true
+rdkafka = { version = "0.36.2", features = ["gssapi", "tracing"] }
 receipts.workspace = true
 reqwest.workspace = true
 secp256k1.workspace = true
