+++ gateway-framework/src/reporting/mod.rs
@@ -1,0 +1,7 @@
+mod kafka;
+mod logging;
+mod metrics;
+
+pub use kafka::{EventFilterFn, EventHandlerFn, KafkaClient};
+pub use logging::{error_log, init, LoggingOptions};
+pub use metrics::{with_metric, METRICS};
